requirements.txt
fastapi
uvicorn
python-multipart
python-docx
requests
jinja2
python-dotenv
